src/modeller/requirements.txt
azure-identity
azure-storage-blob
azure-storage-queue
numpy
pandas
tensorflow
keras
pillow
